working_file_raycluster/requirements.txt
ray[serve]==2.10.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
tritonclient[http]==2.41.0
Pillow==10.0.0
numpy==1.24.3
python-multipart